backend/requirements.txt
fastapi
uvicorn
resend
python-dotenv
supabase
pydantic[email]